chore: Extract loaded key and x509 cert starlark types into separate classes.
